Merge repo

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-common/src/test/java/org/apache/hadoop/yarn/util/resource/TestResources.java
@@ -24,16 +24,16 @@ import static org.junit.Assert.assertTrue;
 
 public class TestResources {
   
-  public Resource createResource(int memory, int vCores) {
+  public Resource createResource(long memory, int vCores) {
     return Resource.newInstance(memory, vCores);
   }
 
   @Test(timeout=1000)
   public void testCompareToWithUnboundedResource() {
     assertTrue(Resources.unbounded().compareTo(
-            createResource(Integer.MAX_VALUE, Integer.MAX_VALUE)) == 0);
+            createResource(Long.MAX_VALUE, Integer.MAX_VALUE)) == 0);
     assertTrue(Resources.unbounded().compareTo(
-        createResource(Integer.MAX_VALUE, 0)) > 0);
+        createResource(Long.MAX_VALUE, 0)) > 0);
     assertTrue(Resources.unbounded().compareTo(
         createResource(0, Integer.MAX_VALUE)) > 0);
   }
